fix: set the az account with the right subscription id to load the k8s cluster config (#5)
This commit adds a fix to set the az account with the right subscription id before loading the right k8s config which fixes the original error of cluster connection. The account subscription required to connect to the cluster is not the default one which leads to the cluster connection error.